ShapeFunctions

Tardigrade.ShapeFunctions — Module

ShapeFunctions

A module for working with FEM shape functions.

Tardigrade.ShapeFunctions.AbstractShapeFunction — Type

AbstractShapeFunction

Parent type of all shape functions.

Tardigrade.ShapeFunctions.LagrangeShapeFunction — Type

LagrangeShapeFunction

Container for shape function methods on the master element.

Tardigrade.ShapeFunctions.LagrangeShapeFunction — Method

LagrangeShapeFunction(element_type::String)

Init method for LagrangeShapeFunction.

Tardigrade.ShapeFunctions.J — Method

J(∇φ_ξ::Matrix{T}, X::S) where {T <: Real, S}

Mapping Jacobian. Written to except views of coordinates.

Tardigrade.ShapeFunctions.detJ — Method

detJ(∇φ_ξ::Matrix{T}, X::S) where {T <: Real, S}

Determinant of mapping Jacobian. Written to except views of coordinates

Tardigrade.ShapeFunctions.map_∇φ_ξ — Method

map_∇φ_ξ(∇φ_ξ::Matrix{T}, X::S) where {T <: Real, S}

Method to map master element shape function grads to physical element. Written to except views.

Tardigrade.ShapeFunctions.quad4_shape_function_gradients! — Method

quad4_shape_function_gradients!

Implementation for quad4 element.

Tardigrade.ShapeFunctions.quad4_shape_function_gradients — Method

quad4_shape_function_gradients(ξ::Vector{T}) where {T <: Real}

Implementation for quad4 element.

Tardigrade.ShapeFunctions.quad4_shape_function_values! — Method

quad4_shape_function_values!(φ::Vector{T}, ξ::Vector{T}) where {T <: Real}

Implementation for quad4 element with pre-allocation.

Tardigrade.ShapeFunctions.quad4_shape_function_values — Method

quad4_shape_function_values(ξ::Vector{T}) where {T <: Real}

Implementation for quad4 element.